Split buildrunner/job/watcher into separate packages.
