+++ src/gbserver/api/builds.py
@@ -32,7 +32,7 @@ from gbserver.api.utils import (
     split_tags,
 )
 from gbserver.buildwatcher.buildwatcher import BuildWatcher
-from gbserver.buildwatcher.validation import BuildValidation
+from gbserver.buildrunner.validation import BuildValidation
 from gbserver.storage.artifact_registration import ArtifactRegistration
 from gbserver.storage.build_storage import IStoredBuildStorage
 from gbserver.storage.singleton_storage import SingletonAdminStorage, get_admin_storage
